Parallel

+++ main/do_video.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
-gifname=$(date +"%b%d-%H%M")
-w_dir=$(pwd)
+echo "=================VIDEO=================" 1>&2
+myUserName=$(whoami)
+gifnamex=$(date +%b%d_%k%M_%S)
+gifname=$(echo ${gifnamex//[[:blank:]]/})
+w_dir=$1
+dir=$(pwd)
 FPS=$(sed '1q;d' "$w_dir/gif_settings.txt")
 FPS=${FPS:6}
 put_MARK=$(sed '10q;d' "$w_dir/gif_settings.txt")
@@ -10,38 +14,32 @@ MARK=${MARK:6}
 QUALITY=$(sed '5q;d' "$w_dir/gif_settings.txt")
 QUALITY=${QUALITY:8}
 
-if [ $QUALITY -lt $((100)) ]; then
-    mogrify -resize "$QUALITY%" "$(pwd)/images/*.JPG"
+GIFDONE=$(bash "$dir/file_exist.sh" "$w_dir" "mp4")
+
+if [[ $GIFDONE -ne 1 ]]; then
+  #insert calibration script here 
+  #read and apply rotation
+  bash backImg.sh "$w_dir"  
+  bash net_gif.sh "$w_dir" "$gifname"
+  if [ $QUALITY -lt $((100)) ]; then
+      mogrify -resize $QUALITY"%" $w_dir/images/*.JPG
+      sleep 2s
+  fi
+
+  echo $gifname.mp4 > $w_dir/gif_name.txt
+
+  echo '====================MP4============================'
+
+  ffmpeg -framerate $FPS -pattern_type glob -i "$w_dir/images/*.JPG" -c:v libx264 -pix_fmt yuv420p "$w_dir/$gifname.mp4"
+  #ffmpeg -framerate $FPS -start_number $((11)) -i $w_dir/images/image-%d.JPG -c:v libx264 -pix_fmt yuv420p "$w_dir/$gifname.mp4"
+
+  if [ $put_MARK -eq $((1)) ]; then
+      echo '===================WATERMARK================================'
+      #echo $MARK
+      ffmpeg -i $w_dir/$gifname.mp4 -i $MARK -filter_complex "overlay=x=(main_w-overlay_w)/2:y=(main_h-overlay_h)/2" -c:v libx264 -pix_fmt yuv420p $w_dir/b$gifname.mp4
+      rm $w_dir/$gifname.mp4
+      mv $w_dir/b$gifname.mp4 $w_dir/$gifname.mp4
+  fi
+  
+  cp "$w_dir/$gifname.mp4" "/home/$myUserName/Pictures/$gifname.mp4"
 fi
-
-echo $gifname.mp4 >gif_name.txt
-
-echo '====================MP4============================'
-
-
-ffmpeg -framerate $FPS -pattern_type glob -i $w_dir'/images/*.JPG' -c:v libx264 -pix_fmt yuv420p $w_dir/$gifname.mp4
-
-#sleep 2
-
-if [ $put_MARK -eq $((1)) ]; then
-    echo '===================WATERMARK================================'
-    echo $MARK
-    ffmpeg -vcodec libx264 -i $w_dir/$gifname.mp4 -i $MARK -filter_complex "overlay=x=(main_w-overlay_w)/2:y=(main_h-overlay_h)/2" b$gifname.mp4
-    rm $w_dir/$gifname.mp4
-    mv $w_dir/b$gifname.mp4 $w_dir/$gifname.mp4
-fi
-
-file="$w_dir/results.txt"
-box_ready=false
-while [[ "$box_ready" == false ]]; do
-    echo -n "o" >/dev/ttyACM0
-    timeout 0.1 cat </dev/ttyACM0 | tee $file
-    line=$(sed '3q;d' $file)
-    if [[ "$line" == Ready* ]]; then
-        box_ready=true
-    else
-        box_ready=false
-    fi
-done
-echo -e "Box Ready\n"
-mv "$w_dir/$gifname.mp4" "/home/onikom/Pictures/$gifname.mp4"
